move to internal and merge selector types
